Need to complete ConfirmAppointment; Need to figure out how to get the Agent, Date, and Time that were chosen by the customer
Additional comments by Youssef:
when creating an appointment object, u need the cart ye(a list of stores). so where is this list stored? in the main menu when u click those buttons. So u need to pass it from the main menu to the select agent(or u could just send it back to the leasing office directly). then from there u select an agent and send it to the leasing office along with the list of stores

u create the appointment object

and u use that object to display all the info for the confirmation message